+++ examples/counter/lib/main.dart
@@ -30,7 +30,7 @@ class Home extends StatelessWidget {
       body: Center(
         // Consumer is a widget that allows you reading providers.
         // You could also use the hook "useProvider" if you uses flutter_hooks
-        child: Consumer((context, watch) {
+        child: Consumer(builder: (context, watch, _) {
           final count = watch(counterProvider).state;
           return Text('$count');
         }),