+++ packages/flutter_riverpod/test/auto_dispose_change_notifier_provider_test.dart
@@ -94,7 +94,7 @@ void main() {
 
     await tester.pumpWidget(
       ProviderScope(
-        child: Consumer((c, watch) {
+        child: Consumer(builder: (c, watch, _) {
           return Text(
             watch(provider).count.toString(),
             textDirection: TextDirection.ltr,