+++ packages/flutter_riverpod/test/provider_test.dart
@@ -42,7 +42,7 @@ void main() {
 
     await tester.pumpWidget(
       ProviderScope(
-        child: Consumer((c, watch) {
+        child: Consumer(builder: (c, watch, _) {
           return Text(
             watch(provider).toString(),
             textDirection: TextDirection.ltr,
@@ -65,7 +65,7 @@ void main() {
 
     await tester.pumpWidget(
       ProviderScope(
-        child: Consumer((c, watch) {
+        child: Consumer(builder: (c, watch, _) {
           return Text(
             watch(provider).toString(),
             textDirection: TextDirection.ltr,
@@ -95,7 +95,7 @@ void main() {
 
     await tester.pumpWidget(
       ProviderScope(
-        child: Consumer((c, watch) {
+        child: Consumer(builder: (c, watch, _) {
           return Text(
             watch(provider).toString(),
             textDirection: TextDirection.ltr,
@@ -137,7 +137,7 @@ void main() {
 
     await tester.pumpWidget(
       ProviderScope(
-        child: Consumer((c, watch) {
+        child: Consumer(builder: (c, watch, _) {
           return Text(
             watch(provider).toString(),
             textDirection: TextDirection.ltr,
@@ -153,7 +153,7 @@ void main() {
   testWidgets('override updates rebuild dependents with new value',
       (tester) async {
     final provider = Provider((_) => 0);
-    final child = Consumer((c, watch) {
+    final child = Consumer(builder: (c, watch, _) {
       return Text(
         watch(provider).toString(),
         textDirection: TextDirection.ltr,
@@ -212,7 +212,7 @@ void main() {
             }),
           ),
         ],
-        child: Consumer((c, watch) {
+        child: Consumer(builder: (c, watch, _) {
           return Text(
             watch(provider2).toString(),
             textDirection: TextDirection.ltr,
@@ -237,7 +237,7 @@ void main() {
         overrides: [
           provider.overrideWithProvider(Provider((_) => 1)),
         ],
-        child: Consumer((c, watch) {
+        child: Consumer(builder: (c, watch, _) {
           return Text(watch(provider1), textDirection: TextDirection.ltr);
         }),
       ),
@@ -261,7 +261,7 @@ void main() {
 
     await tester.pumpWidget(
       ProviderScope(
-        child: Consumer((c, watch) {
+        child: Consumer(builder: (c, watch, _) {
           return Text(
             watch(forth).toString(),
             textDirection: TextDirection.ltr,
@@ -290,7 +290,7 @@ void main() {
         overrides: [
           first.overrideWithProvider(Provider((_) => 42)),
         ],
-        child: Consumer((c, watch) {
+        child: Consumer(builder: (c, watch, _) {
           return Text(
             watch(forth).toString(),
             textDirection: TextDirection.ltr,
@@ -319,7 +319,7 @@ void main() {
         overrides: [
           second.overrideWithProvider(Provider((_) => 0)),
         ],
-        child: Consumer((c, watch) {
+        child: Consumer(builder: (c, watch, _) {
           return Text(
             watch(forth).toString(),
             textDirection: TextDirection.ltr,
@@ -346,7 +346,7 @@ void main() {
 
     await tester.pumpWidget(
       ProviderScope(
-        child: Consumer((c, watch) {
+        child: Consumer(builder: (c, watch, _) {
           return Text(
             watch(provider1).toString(),
             textDirection: TextDirection.ltr,