+++ tools/generate_providers/bin/generate_providers.dart
@@ -126,12 +126,12 @@ const _familyDoc = r'''
 ///   // ...
 ///
 ///   @override
-///   Widget build(BuildContext context) {
+///   Widget build(BuildContext context, ScopedReader watch) {
 ///     final locale = Localizations.localeOf(context);
 ///
 ///     // Obtains the title based on the current Locale.
 ///     // Will automatically update the title when the Locale changes.
-///     final title = useProvider(titleFamily(locale));
+///     final title = watch(titleFamily(locale));
 ///
 ///     return Text(title);
 ///   }
@@ -148,14 +148,14 @@ const _familyDoc = r'''
 ///   // ...
 ///
 ///   @override
-///   Widget build(BuildContext context) {
+///   Widget build(BuildContext context, ScopedReader watch) {
 ///     int userId; // Read the user ID from somewhere
 ///
 ///     // Read and potentially fetch the user with id `userId`.
 ///     // When `userId` changes, this will automatically update the UI
 ///     // Similarly, if two widgets tries to read `userFamily` with the same `userId`
 ///     // then the user will be fetched only once.
-///     final user = useProvider(userFamily(userId));
+///     final user = watch(userFamily(userId));
 ///
 ///     return user.when(
 ///       data: (user) => Text(user.name),
@@ -193,9 +193,9 @@ const _familyDoc = r'''
 /// The usual:
 /// 
 /// ```dart
-/// Widget build(BuildContext) {
+/// Widget build(BuildContext, ScopedReader watch) {
 ///   // Error – messagesFamily is not a provider
-///   final response = useProvider(messagesFamily);
+///   final response = watch(messagesFamily);
 /// }
 /// ```
 ///
@@ -203,8 +203,8 @@ const _familyDoc = r'''
 /// Instead, we need to pass a parameter to `messagesFamily`:
 ///
 /// ```dart
-/// Widget build(BuildContext) {
-///   final response = useProvider(messagesFamily('id'));
+/// Widget build(BuildContext, ScopedReader watch) {
+///   final response = watch(messagesFamily('id'));
 /// }
 /// ```
 ///
@@ -214,9 +214,9 @@ const _familyDoc = r'''
 ///
 /// ```dart
 /// @override
-/// Widget build(BuildContext context) {
-///   final frenchTitle = useProvider(titleFamily(const Locale('fr')));
-///   final englishTitle = useProvider(titleFamily(const Locale('en')));
+/// Widget build(BuildContext context, ScopedReader watch) {
+///   final frenchTitle = watch(titleFamily(const Locale('fr')));
+///   final englishTitle = watch(titleFamily(const Locale('en')));
 ///
 ///   return Text('fr: $frenchTitle en: $englishTitle');
 /// }
@@ -272,11 +272,11 @@ const _familyDoc = r'''
 /// })
 ///
 /// @override
-/// Widget build(BuildContext context) {
+/// Widget build(BuildContext context, ScopedReader watch) {
 ///   int userId; // Read the user ID from somewhere
 ///   final locale = Localizations.localeOf(context);
 ///
-///   final something = useProvider(
+///   final something = watch(
 ///     exampleProvider(MyParameter(userId: userId, locale: locale)),
 ///   );
 ///